refactor: Column과 점수 계산 클래스 분리
