+++ src/main/java/org/example/chess/Board.java
@@ -10,41 +10,6 @@ import org.example.pieces.Piece.Type;
 
 public class Board {
 
-    public class Column {
-
-        // 자신의 col과 row길이를 생성시 받는다.
-        private final char colIndex;
-        private final int rowLen;
-        private final List<Piece> pieces;
-
-        public Column(char colIndex, int rowLen) {
-            this.colIndex = colIndex;
-            this.rowLen = rowLen;
-            this.pieces = createSizeList(rowLen);
-        }
-
-        private List<Piece> createSizeList(int boardSize) {
-            List<Piece> ret = new ArrayList<>();
-            for (int i = 0; i < boardSize; i++) {
-                ret.add(Piece.createNoColorPiece());
-            }
-            return ret;
-        }
-
-        private void setPiece(Piece piece, int rowNumber) {
-            pieces.set(rowNumber - 1, piece);
-        }
-
-        public Piece getPiece(int index) {
-            // 1 을 빼줘야한다.
-            return pieces.get(index - 1);
-        }
-
-        public List<Piece> getPieces() {
-            return pieces;
-        }
-    }
-
     private final List<Column> columns;
 
     private final int BOARD_SIZE = 8;
@@ -146,7 +111,7 @@ public class Board {
 
     public int nonEmptyPiece() {
         return columns.stream()
-            .mapToInt(column -> (int) column.pieces.stream()
+            .mapToInt(column -> (int) column.getPieces().stream()
                 .filter(piece -> piece.isBlack() || piece.isWhite())
                 .count())
             .sum();
@@ -154,7 +119,7 @@ public class Board {
 
     public int countByQuery(Color color, Type type) {
         return columns.stream()
-            .mapToInt(column -> (int) column.pieces.stream()
+            .mapToInt(column -> (int) column.getPieces().stream()
                 .filter(piece -> piece.isSameColor(color) && piece.isSameType(type))
                 .count())
             .sum();
@@ -174,7 +139,7 @@ public class Board {
 
     public List<Piece> sortByPoint(Color color) {
         return columns.stream()
-            .flatMap(column -> column.pieces.stream())
+            .flatMap(column -> column.getPieces().stream())
             .filter(piece -> piece.isSameColor(color))
             .sorted((p1, p2) -> Double.compare(p2.getPoint(), p1.getPoint()))
             .toList();
